trupa_bot/requirements.txt
python-telegram-bot==13.15
stable-baselines3
gymnasium
ta
yfinance
pandas
numpy
requests
